Add bookmark-user ownership

+++ src/bookmarks/bookmarks.service.ts
@@ -3,6 +3,7 @@ import { CreateBookmarkDto } from './dto/create-bookmark.dto';
 import { BookmarkRepository } from './bookmark.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Bookmark } from './bookmark.entity';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class BookmarksService {
@@ -11,12 +12,19 @@ export class BookmarksService {
     private bookmarkRepository: BookmarkRepository,
   ) {}
 
-  async getAllBookmarks(): Promise<Bookmark[]> {
-    return this.bookmarkRepository.getBookmarks();
+  async getAllBookmarks(
+    user: User,
+  ): Promise<Bookmark[]> {
+    return this.bookmarkRepository.getBookmarks(user);
   }
 
-  async getBookmarkById(id: number): Promise<Bookmark> {
-    const found = await this.bookmarkRepository.findOne(id);
+  async getBookmarkById(
+    id: number,
+    user: User,
+  ): Promise<Bookmark> {
+    const found = await this.bookmarkRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException();
     }
@@ -24,12 +32,18 @@ export class BookmarksService {
     return found;
   }
 
-  async createBookmark(createBookmarkDto: CreateBookmarkDto): Promise<Bookmark> {
-    return this.bookmarkRepository.createBookmark(createBookmarkDto);
+  async createBookmark(
+    createBookmarkDto: CreateBookmarkDto,
+    user: User,
+  ): Promise<Bookmark> {
+    return this.bookmarkRepository.createBookmark(createBookmarkDto, user);
   }
 
-  async deleteBookmark(id: number): Promise<void> {
-    const result = await this.bookmarkRepository.delete(id);
+  async deleteBookmark(
+    id: number,
+    user: User,
+  ): Promise<void> {
+    const result = await this.bookmarkRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException();
     }
